cohort-1/track-1/firstresponse-ai/requirements.txt
google-genai>=1.0
flask>=3.0
flask-cors>=4.0
gunicorn>=22.0
pytest>=8.0
